ENH: Huge refactor. Move things to logical modules
* fit.py is the main entry point
* paramselect.py is where all of the fitting happens. This is the core.
* core_utils.py contains specialized utilities for ESPEI.
* utils.py are utilities with reuse potential outside of ESPEI.
* plot.py holds plotting functions
* Kill dead code
