+++ espei/core_utils.py
@@ -1,0 +1,178 @@
+"""
+Module for handling data
+"""
+import itertools
+import operator
+from functools import reduce
+import copy
+import numpy as np
+import tinydb
+from pycalphad import variables as v
+
+
+def get_data(comps, phase_name, configuration, symmetry, datasets, prop):
+    configuration = list(configuration)
+    desired_data = datasets.search((tinydb.where('output').test(lambda x: x in prop)) &
+                                   (tinydb.where('components').test(lambda x: set(x).issubset(comps))) &
+                                   (tinydb.where('solver').test(symmetry_filter, configuration, symmetry)) &
+                                   (tinydb.where('phases') == [phase_name]))
+    # This seems to be necessary because the 'values' member does not modify 'datasets'
+    # But everything else does!
+    desired_data = copy.deepcopy(desired_data)
+    #if len(desired_data) == 0:
+    #    raise ValueError('No datasets for the system of interest containing {} were in \'datasets\''.format(prop))
+
+    def recursive_zip(a, b):
+        if isinstance(a, (list, tuple)) and isinstance(b, (list, tuple)):
+            return list(recursive_zip(x, y) for x, y in zip(a, b))
+        else:
+            return list(zip(a, b))
+
+    for idx, data in enumerate(desired_data):
+        # Filter output values to only contain data for matching sublattice configurations
+        matching_configs = np.array([(canonicalize(sblconf, symmetry) == canonicalize(configuration, symmetry))
+                                     for sblconf in data['solver']['sublattice_configurations']])
+        matching_configs = np.arange(len(data['solver']['sublattice_configurations']))[matching_configs]
+        # Rewrite output values with filtered data
+        desired_data[idx]['values'] = np.array(data['values'], dtype=np.float)[..., matching_configs]
+        desired_data[idx]['solver']['sublattice_configurations'] = list_to_tuple(np.array(data['solver']['sublattice_configurations'],
+                                                                                          dtype=np.object)[matching_configs].tolist())
+        try:
+            desired_data[idx]['solver']['sublattice_occupancies'] = np.array(data['solver']['sublattice_occupancies'],
+                                                                             dtype=np.object)[matching_configs].tolist()
+        except KeyError:
+            pass
+        # Filter out temperatures below 298.15 K (for now, until better refstates exist)
+        temp_filter = np.atleast_1d(data['conditions']['T']) >= 298.15
+        desired_data[idx]['conditions']['T'] = np.atleast_1d(data['conditions']['T'])[temp_filter]
+        # Don't use data['values'] because we rewrote it above; not sure what 'data' references now
+        desired_data[idx]['values'] = desired_data[idx]['values'][..., temp_filter, :]
+    return desired_data
+
+
+def get_samples(desired_data):
+    all_samples = []
+    for data in desired_data:
+        temperatures = np.atleast_1d(data['conditions']['T'])
+        num_configs = np.array(data['solver'].get('sublattice_configurations'), dtype=np.object).shape[0]
+        site_fractions = data['solver'].get('sublattice_occupancies', [[1]] * num_configs)
+        site_fraction_product = [reduce(operator.mul, list(itertools.chain(*[np.atleast_1d(f) for f in fracs])), 1)
+                                 for fracs in site_fractions]
+        # TODO: Subtle sorting bug here, if the interactions aren't already in sorted order...
+        interaction_product = []
+        for fracs in site_fractions:
+            interaction_product.append(float(reduce(operator.mul,
+                                                    [f[0] - f[1] for f in fracs if isinstance(f, list) and len(f) == 2],
+                                                    1)))
+        if len(interaction_product) == 0:
+            interaction_product = [0]
+        comp_features = zip(site_fraction_product, interaction_product)
+        all_samples.extend(list(itertools.product(temperatures, comp_features)))
+    return all_samples
+
+
+def canonicalize(configuration, equivalent_sublattices):
+    """
+    Sort a sequence with symmetry. This routine gives the sequence
+    a deterministic ordering while respecting symmetry.
+
+    Parameters
+    ==========
+    configuration : list
+        Sublattice configuration to sort.
+    equivalent_sublattices : set of set of int
+        Indices of 'configuration' which should be equivalent by symmetry, i.e.,
+        [[0, 4], [1, 2, 3]] means permuting elements 0 and 4, or 1, 2 and 3, respectively,
+        has no effect on the equivalence of the sequence.
+
+    Returns
+    =======
+    canonicalized : tuple
+    """
+    canonicalized = list(configuration)
+    if equivalent_sublattices is not None:
+        for subl in equivalent_sublattices:
+            subgroup = sorted([configuration[idx] for idx in sorted(subl)], key=canonical_sort_key)
+            for subl_idx, conf_idx in enumerate(sorted(subl)):
+                if isinstance(subgroup[subl_idx], list):
+                    canonicalized[conf_idx] = tuple(subgroup[subl_idx])
+                else:
+                    canonicalized[conf_idx] = subgroup[subl_idx]
+
+    return list_to_tuple(canonicalized)
+
+
+def symmetry_filter(x, config, symmetry):
+    if x['mode'] == 'manual':
+        if len(config) != len(x['sublattice_configurations'][0]):
+            return False
+        # If even one matches, it's a match
+        # We do more filtering downstream
+        for data_config in x['sublattice_configurations']:
+            if canonicalize(config, symmetry) == canonicalize(data_config, symmetry):
+                return True
+    return False
+
+
+def canonical_sort_key(x):
+    """
+    Wrap strings in tuples so they'll sort.
+
+    Parameters
+    ==========
+    x : sequence
+    """
+    return [tuple(i) if isinstance(i, (tuple, list)) else (i,) for i in x]
+
+
+def list_to_tuple(x):
+    def _tuplify(y):
+        if isinstance(y, list) or isinstance(y, tuple):
+            return tuple(_tuplify(i) if isinstance(i, (list, tuple)) else i for i in y)
+        else:
+            return y
+    return tuple(map(_tuplify, x))
+
+
+def endmembers_from_interaction(configuration):
+    config = []
+    for c in configuration:
+        if isinstance(c, (list, tuple)):
+            config.append(c)
+        else:
+            config.append([c])
+    return list(itertools.product(*[tuple(c) for c in config]))
+
+
+def build_sitefractions(phase_name, sublattice_configurations, sublattice_occupancies):
+    """
+    Convert nested lists of sublattice configurations and occupancies to a list of dictionaries.
+    The dictionaries map SiteFraction symbols to occupancy values. Note that zero occupancy
+    site fractions will need to be added separately since the total degrees of freedom aren't
+    known in this function.
+    :param phase_name:
+    :param sublattice_configurations:
+    :param sublattice_occupancies:
+    :return:
+    """
+    result = []
+    for config, occ in zip(sublattice_configurations, sublattice_occupancies):
+        sitefracs = {}
+        config = [[c] if not isinstance(c, (list, tuple)) else c for c in config]
+        occ = [[o] if not isinstance(o, (list, tuple)) else o for o in occ]
+        if len(config) != len(occ):
+            raise ValueError('Sublattice configuration length differs from occupancies')
+        for sublattice_idx in range(len(config)):
+            if isinstance(config[sublattice_idx], (list, tuple)) != isinstance(occ[sublattice_idx], (list, tuple)):
+                raise ValueError('Sublattice configuration type differs from occupancies')
+            if not isinstance(config[sublattice_idx], (list, tuple)):
+                # This sublattice is fully occupied by one component
+                sitefracs[v.SiteFraction(phase_name, sublattice_idx, config[sublattice_idx])] = occ[sublattice_idx]
+            else:
+                # This sublattice is occupied by multiple elements
+                if len(config[sublattice_idx]) != len(occ[sublattice_idx]):
+                    raise ValueError('Length mismatch in sublattice configuration')
+                for comp, val in zip(config[sublattice_idx], occ[sublattice_idx]):
+                    sitefracs[v.SiteFraction(phase_name, sublattice_idx, comp)] = val
+        result.append(sitefracs)
+    return result
